Undo data: Fill missing currency in omitted data from API
When an entry got submitted via API with an account line with only
account name, and the amount & currency omitted, when undo data didn't
contain even the default currency.  It caused minor usability problems
when amending, namely the JS didn't complete the omitted value as a
placeholder and more importantly the form couldn't be submitted as
this field was mandatory (TODO: why is it mandatory?).