More robust replacement rules
Among other things, it's now possible to use the regex capture groups
in the replacement strings.

Previously whether the rule is applied to a field was decided by the
presence of the replacement, not the pattern, so it was not possible
to clear a field.  It's now decided by the presence of the pattern, so
an empty replacement clears the field if this field is present in the
condition.  It's still possible to match with /(regex)/ and replace
with \1 to leave the field untouched.